Make debian/rules get-orig-source smarter
Instead of running uscan, check the version first. If it is a git
snapshot, then create a make dist tarball ourselves. Otherwise call
uscan as before.

Because this is slightly more complicated than what is convenient to do
in make, delegate the task to a shell script under debian/tools.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>